Rename `EscapedString` by `EscapeSequence`
